Update folder structure

+++ dev.sh
@@ -1,3 +1,11 @@
-clear
+#!/bin/sh
+
+if [ "${1}" = "setup" ]; then
+    docker-compose build || exit 1
+    docker-compose run web sh -c "npm ci"
+    docker-compose run server sh -c "npm ci && npm run build:clean"
+    docker-compose run tests sh -c "npm ci && npm run build:clean"
+fi
+
 docker-compose down
-docker-compose up -d
+docker-compose up --abort-on-container-exit
